more doubles to shut up clang

+++ valhalla/meili/match_result.h
@@ -68,16 +68,16 @@ struct MatchResult {
 
 struct EdgeSegment {
   EdgeSegment(baldr::GraphId the_edgeid,
-              float the_source = 0.f,
-              float the_target = 1.f,
+              double the_source = 0.f,
+              double the_target = 1.f,
               int the_first_match_idx = -1,
               int the_last_match_idx = -1,
               bool disconnect = false,
               int restriction_idx = -1);
 
   baldr::GraphId edgeid;
-  float source{0.f};
-  float target{1.f};
+  double source{0.f};
+  double target{1.f};
   int first_match_idx{-1};
   int last_match_idx{-1};
   int restriction_idx{-1};
